+++ src/Okanshi/Monitor.fs
@@ -27,21 +27,21 @@ type OkanshiMonitor private () =
             tagDictionary.Clear()
             value |> Seq.iter (fun x -> tagDictionary.TryAdd(x, byte 0) |> ignore)
     
-    /// Get or add a BasicCounter
-    static member BasicCounter(name : string) = OkanshiMonitor.BasicCounter(name, [||])
+    /// Get or add a CumulativeCounter
+    static member CumulativeCounter(name : string) = OkanshiMonitor.CumulativeCounter(name, [||])
     
-    /// Get or add a BasicCounter with custom tags
-    static member BasicCounter(name : string, tags : Tag array) = 
+    /// Get or add a CumulativeCounter with custom tags
+    static member CumulativeCounter(name : string, tags : Tag array) = 
         let config = MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags)
-        monitorRegistry.GetOrAdd(config, fun x -> new BasicCounter(x))
+        monitorRegistry.GetOrAdd(config, fun x -> new CumulativeCounter(x))
     
     /// Get or add a PeakRateCounter
-    static member PeakCounter(name) = OkanshiMonitor.PeakCounter(name, [||])
+    static member Counter(name) = OkanshiMonitor.Counter(name, [||])
     
     /// Get or add a PeakRateCounter with custom tags
-    static member PeakCounter(name : string, tags : Tag array) = 
+    static member Counter(name : string, tags : Tag array) = 
         let config = MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags)
-        monitorRegistry.GetOrAdd(config, fun x -> new PeakCounter(x))
+        monitorRegistry.GetOrAdd(config, fun x -> new Counter(x))
     
     /// Get or add a DoubleCounter
     static member DoubleCounter(name) = OkanshiMonitor.DoubleCounter(name, [||])
@@ -51,13 +51,13 @@ type OkanshiMonitor private () =
         let config = MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags)
         monitorRegistry.GetOrAdd(config, fun x -> new DoubleCounter(x))
     
-    /// Get or add a BasicGauge
-    static member BasicGauge(name : string, getValue : Func<'T>) = OkanshiMonitor.BasicGauge(name, getValue, [||])
+    /// Get or add a Gauge
+    static member Gauge(name : string, getValue : Func<'T>) = OkanshiMonitor.Gauge(name, getValue, [||])
     
-    /// Get or add a BasicGauge with custom tags
-    static member BasicGauge<'T>(name : string, getValue : Func<'T>, tags : Tag array) = 
+    /// Get or add a Gauge with custom tags
+    static member Gauge<'T>(name : string, getValue : Func<'T>, tags : Tag array) = 
         let config = MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags)
-        monitorRegistry.GetOrAdd(config, fun x -> new BasicGauge<'T>(x, getValue))
+        monitorRegistry.GetOrAdd(config, fun x -> new Gauge<'T>(x, getValue))
     
     /// Get or add a MaxGauge 
     static member MaxGauge(name : string) = OkanshiMonitor.MaxGauge(name, [||])
@@ -107,13 +107,13 @@ type OkanshiMonitor private () =
         let config = MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags)
         monitorRegistry.GetOrAdd(config, fun x -> new DecimalGauge(x))
     
-    /// Get or add a BasicTimer, with a step size of 1 minute
-    static member BasicTimer(name) = OkanshiMonitor.BasicTimer(name, [||])
+    /// Get or add a Timer, with a step size of 1 minute
+    static member Timer(name) = OkanshiMonitor.Timer(name, [||])
     
-    /// Get or add a BasicTimer with custom tags
-    static member BasicTimer(name : string, tags : Tag array) = 
+    /// Get or add a Timer with custom tags
+    static member Timer(name : string, tags : Tag array) = 
         let config = MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags)
-        monitorRegistry.GetOrAdd(config, fun x -> new BasicTimer(x))
+        monitorRegistry.GetOrAdd(config, fun x -> new Timer(x))
     
 #if NET45
     /// Get or add a performance counter monitor